Renames auto to single
