--- example-viewmodel/src/main/java/com/mlykotom/exampleviewmodel/auto/AutoValidationView.java
@@ -1,8 +1,0 @@
-package com.mlykotom.exampleviewmodel.auto;
-
-import eu.inloop.viewmodel.IView;
-
-
-public interface AutoValidationView extends IView {
-	void onSubmitClicked();
-}
