+++ example-viewmodel/src/main/java/com/mlykotom/exampleviewmodel/single/SingleValidationView.java
@@ -1,0 +1,8 @@
+package com.mlykotom.exampleviewmodel.single;
+
+import eu.inloop.viewmodel.IView;
+
+
+public interface SingleValidationView extends IView {
+	void onSubmitClicked();
+}
